0.5.3 update package + separate autocomplete and signature
