+++ package.json
@@ -399,7 +399,7 @@
     "lru-cache": "^7.14.1",
     "miniscript-meta": "~1.2.2",
     "miniscript-textmate": "~0.2.1",
-    "miniscript-type-analyzer": "~0.1.9",
+    "miniscript-type-analyzer": "~0.1.11",
     "text-encoder-lite": "^2.0.0"
   },
   "browserslist": "> 0.25%, not dead"
