+++ src/autocomplete.ts
@@ -1,17 +1,8 @@
 import {
   ASTBase,
-  ASTCallExpression,
   ASTIndexExpression,
   ASTMemberExpression
 } from 'miniscript-core';
-import {
-  SignatureDefinitionFunction,
-  SignatureDefinitionFunctionArg,
-  SignatureDefinitionTypeMeta,
-  SignatureDefinitionBaseType,
-  SignatureDefinition
-} from 'meta-utils';
-import { miniscriptMeta } from 'miniscript-meta';
 import vscode, {
   CancellationToken,
   CompletionContext,
@@ -19,11 +10,7 @@ import vscode, {
   CompletionItemKind,
   CompletionList,
   ExtensionContext,
-  ParameterInformation,
   Position,
-  SignatureHelp,
-  SignatureHelpContext,
-  SignatureInformation,
   TextDocument
 } from 'vscode';
 
@@ -162,99 +149,4 @@ export function activate(_context: ExtensionContext) {
       return new CompletionList(completionItems);
     }
   }, '.');
-
-  vscode.languages.registerSignatureHelpProvider(
-    'miniscript',
-    {
-      async provideSignatureHelp(
-        document: TextDocument,
-        position: Position,
-        _token: CancellationToken,
-        _ctx: SignatureHelpContext
-      ): Promise<SignatureHelp> {
-        documentParseQueue.refresh(document);
-        const helper = new LookupHelper(document);
-        const astResult = helper.lookupAST(position);
-
-        if (!astResult) {
-          return;
-        }
-
-        // filter out root call expression for signature
-        let rootCallExpression: ASTCallExpression | undefined;
-
-        if (astResult.closest.type === 'CallExpression') {
-          rootCallExpression = astResult.closest as ASTCallExpression;
-        } else {
-          for (let index = astResult.outer.length - 1; index >= 0; index--) {
-            const current = astResult.outer[index];
-
-            if (current.type === 'CallExpression') {
-              rootCallExpression = current as ASTCallExpression;
-              break;
-            }
-          }
-        }
-
-        if (!rootCallExpression) {
-          return;
-        }
-
-        const root = helper.lookupScope(astResult.closest);
-        const item = await helper.lookupTypeInfo({
-          closest: rootCallExpression.base,
-          outer: root ? [root] : []
-        });
-
-        if (!item || !item.isCallable()) {
-          return;
-        }
-
-        // figure out argument position
-        const astArgs = rootCallExpression.arguments;
-        const selectedIndex = astArgs.findIndex((argItem: ASTBase) => {
-          const leftIndex = argItem.start!.character - 1;
-          const rightIndex = argItem.end!.character;
-
-          return (
-            leftIndex <= position.character && rightIndex >= position.character
-          );
-        });
-
-        const signatureHelp = new SignatureHelp();
-
-        signatureHelp.activeParameter =
-          selectedIndex === -1 ? 0 : selectedIndex;
-        signatureHelp.signatures = [];
-        signatureHelp.activeSignature = 0;
-
-        const fnDef = item.signatureDefinitions.first() as SignatureDefinitionFunction;
-        const args = fnDef.getArguments() || [];
-        const returnValues = fnDef.getReturns().join(' or ') || 'null';
-        const argValues = args
-          .map(
-            (item: SignatureDefinitionFunctionArg) =>
-              `${item.getLabel()}${item.isOptional() ? '?' : ''}: ${item.getTypes().join(' or ')}`
-          )
-          .join(', ');
-        const signatureInfo = new SignatureInformation(
-          `(${SignatureDefinitionBaseType.Function}) ${item.label} (${argValues}): ${returnValues}`
-        );
-        const params: ParameterInformation[] = args.map(
-          (argItem: SignatureDefinitionFunctionArg) => {
-            return new ParameterInformation(
-              `${argItem.getLabel()}${argItem.isOptional() ? '?' : ''}: ${argItem.getTypes().join(' or ')}`
-            );
-          }
-        );
-
-        signatureInfo.parameters = params;
-        signatureHelp.signatures.push(signatureInfo);
-
-        return signatureHelp;
-      }
-    },
-    ',',
-    '('
-  );
 }
