+++ src/extension.ts
@@ -1,6 +1,7 @@
 import vscode, { ExtensionContext } from 'vscode';
 
 import { activate as activateAutocomplete } from './autocomplete';
+import { activate as activateSignature } from './signature';
 import { activate as activateBuild } from './build';
 import { activate as activateDebug } from './debug';
 import { activate as activateDefinition } from './definition';
@@ -25,6 +26,7 @@ export function activate(context: ExtensionContext) {
 
   if (config.get<boolean>('autocomplete')) {
     activateAutocomplete(context);
+    activateSignature(context);
   }
 
   activateDebug(context);
